% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cogaps_loadings_example.r
\docType{data}
\name{cogaps_loadings_example}
\alias{cogaps_loadings_example}
\title{An exaple of loadings of genes in CoGAPS - identified patterns.}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 15176 rows and 8 columns.
}
\source{
\url{https://zenodo.org/records/7709664/files/cogapsresult.Rds}
}
\usage{
cogaps_loadings_example
}
\description{
An exaple of loadings of genes in CoGAPS - identified patterns.
}
\keyword{datasets}
