% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cogaps_loadings_example_markers.r
\docType{data}
\name{cogaps_loadings_example_markers}
\alias{cogaps_loadings_example_markers}
\title{The CoGAPS::patterMarkers result for the example of
loadings of genes in CoGAPS - identified patterns \link{cogaps_loadings_example}.}
\format{
An object of class \code{list} of length 3.
}
\source{
\url{https://zenodo.org/records/7709664/files/cogapsresult.Rds}
}
\usage{
cogaps_loadings_example_markers
}
\description{
The CoGAPS::patterMarkers result for the example of
loadings of genes in CoGAPS - identified patterns \link{cogaps_loadings_example}.
}
\keyword{datasets}
