% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/friends.test.bic.r
\name{friends.test.bic}
\alias{friends.test.bic}
\title{friends.test.bic}
\usage{
friends.test.bic(
  A = NULL,
  prior.to.have.friends = -1,
  max.friends.n = "all",
  .progress = FALSE
)
}
\arguments{
\item{A}{original association matrix}

\item{prior.to.have.friends}{The prior for a row to have friendly columns.}

\item{max.friends.n}{The maximal number of friends for a marker.
A value $n$ means that we filter out a row if it has more
than $n$ friendly columns. 1 means we look only for unique (best) friends.
The string "all" (default) means the same as \code{ncols(A)} value,
do not filter markers by this parameter.}

\item{.progress}{the .progress is passed to \code{purrr} functions
The default is \code{.FALSE}.
If it is not, non-\code{purr} parts also shows progress.}
}
\value{
\code{list}; each element represents a marker, \emph{e.g.},
a matrix row that has friend(s). Each element of the return list
is also a list, one element per friend, and the 2-nd level element
is an integer vector with three numbers, that are:
the marker coordinate (\code{marker}),
the friend coordinate (\code{friend}), and
the the rank of the friend for the marker (\code{rank}).
So, it is list of lists of simple integer vectors, each
vector represents a marker+friend pair,
the inner lists enumerate friends,
the outer (return) list enumerate markers.
}
\description{
We have two sets:T (rows) and C (columns) and
A real matrix A(t,c) that describes the strength of association
between each t and each c; t is an element of T and c is an element of C.
For each t we want to identify whether it is significantly more
relevant for some c's than for the remaining c's.
If it does, those c for which the t is relevant,
are the t's friend. And, the t is the c's marker.
}
\examples{
A <- matrix(
    c(
        10, 6, 7, 8, 9,
        9, 10, 6, 7, 8,
        8, 9, 10, 6, 7,
        7, 8, 9, 10, 6,
        6, 7, 8, 9, 10,
        20, 0, 0, 0, 0
    ),
    nrow = 6, ncol = 5, byrow = TRUE
)
A
friends.test.bic(A, prior.to.have.friends = 0.5)
friends.test.bic(A, prior.to.have.friends = 0.001)
}
