% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unif.ks.test.r
\name{unif.ks.test}
\alias{unif.ks.test}
\title{unif.ks.test}
\usage{
unif.ks.test(ranks, uniform.max = NA, simulate.p.value = FALSE, B = 2000)
}
\arguments{
\item{ranks}{vector of ranks of a tag in different collections, \eqn{1 .. N})}

\item{uniform.max}{the maximal rank in the uniform, default is NA which means
that we take the max(ranks) as the maximal value}

\item{simulate.p.value}{K-S by Monte-Carlo if \code{TRUE};
default is \code{FALSE}, see \code{\link[stats:ks.test]{stats::ks.test()}}}

\item{B}{number of or replicates if \code{simulate.p.value=TRUE}
default is 2000, see \code{\link[stats:ks.test]{stats::ks.test()}}}
}
\value{
p-value for the K-S test comparing the ranks distribution with uniform
}
\description{
returns the Kolomogorov-Smirnov p-value for uniformity of the values.
The values are the ranks of the same row in different columns.
}
\details{
See \link{friends.test} documentation for details.
}
\examples{
example(tag.int.ranks)
ks.p.vals<-apply(TF.ranks,1,"unif.ks.test")
}
