% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unif_ks_test.R
\name{unif_ks_test}
\alias{unif_ks_test}
\title{unif_ks_test}
\usage{
unif_ks_test(ranks, uniform.max = NA, simulate.p.value = FALSE, B = 2000)
}
\arguments{
\item{ranks}{vector of ranks of a row in different columns, \eqn{1 .. N})}

\item{uniform.max}{the maximal rank in the uniform, default is NA which means
that we take the max(ranks) as the maximal value}

\item{simulate.p.value}{K-S by Monte-Carlo if \code{TRUE};
default is \code{FALSE}, see \code{\link[stats:ks.test]{stats::ks.test()}}}

\item{B}{number of or replicates if \code{simulate.p.value=TRUE}
default is 2000, see \code{\link[stats:ks.test]{stats::ks.test()}}}
}
\value{
p-value for the KS test comparing the ranks distribution with uniform
}
\description{
returns the Kolomogorov-Smirnov p-value for uniformity of the values.
The values are the ranks of the same row in different columns.
}
\details{
See \link{friends_test} documentation for details.
}
\examples{
example(row_int_ranks)
ks.p.vals <- apply(TF.ranks, 1, "unif_ks_test")
}
