% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_step_fit.R
\name{best_step_fit}
\alias{best_step_fit}
\title{best_step_fit}
\usage{
best_step_fit(ranks, max.possible.rank)
}
\arguments{
\item{ranks}{integer vector of ranks of a row in different columns}

\item{max.possible.rank}{number of rows, i.e. maximal possible rank}
}
\value{
a list of four values: \cr
\code{step.models} is return from \link{step_fit_ln_likelihoods} call,
which the function starts with
\code{best.step.rank} is the rank value that makes the best step;
it is not obligatory one on the \code{ranks} value.\cr
\code{columns.on.left} is
the vector of the columns on the left of the best step
(including the step value). They are friends of the row.\cr
\code{columns.on.right} is vector of those on the right \cr
\code{population.on.left} is how many ranks are on left of split;
they are friends! \cr
}
\description{
finds the ML-best step model for one row
}
\details{
See \link{friends_test} documentation for details.
}
\examples{
example(row_int_ranks)
step <- best_step_fit(TF.ranks[42, ], genes.no)
}
