% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/friends.test-package.r
\docType{package}
\name{friends.test-package}
\alias{friends.test-package}
\title{friends.test: A package that checks if a partition element in a bipartite
graph has friends in the other partition.}
\description{
We have two sets:T (rows) and C (columns) and
A real matrix A(t,c) that describes the strength of association
between each t and each c; t is an element of T and c is an element of C.
For each t we want to identify whether it is significantly more
relevant for some c's than for the remaining c's.
If it does, those c for which the t is relevant,
are the t's friend. And, the t is the c's marker.
For each tag, we want to identify the collection(s) that specifically prefer(s) the tag.
We say that such a collection is a friend (or the best friend if it is the only) for the tag.
The simplest example: imagine that only one collection pays attention to our tag.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/favorov/best-friends}
  \item Report bugs at \url{https://github.com/favorov/best-friends/issues}
}

}
\author{
\strong{Maintainer}: Alexander Favorov \email{favorov@sensi.org} (\href{https://orcid.org/0000-0003-2595-1091}{ORCID})

Authors:
\itemize{
  \item Alexandra Suvorikova \email{a.suvorikova@gmail.com} [contributor]
  \item Alexey Kroshnin \email{kroshnin@wias-berlin.de} [contributor]
  \item Dmitrijs Lvovs \email{dmitrijs.lvovs@gmail.com} [contributor]
  \item Vera Mukhina \email{atgc@vigg.ru} [contributor]
  \item Andrey A. Mironov \email{mironov@bioinf.fbb.msu.ru} [contributor]
}

}
\keyword{internal}
