% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/friends.test-package.r
\docType{package}
\name{friends.test-package}
\alias{friends.test-package}
\title{friends.test: Rank-Based Method for Feature Selection in Interaction Matrices}
\description{
We have two sets:T (rows) and C (columns) and
A real matrix A(t,c) that describes the strength of association
between each t and each c; t is an element of T and c is an element of C.
For each t we want to identify whether it is significantly more
relevant for some c's than for the remaining c's.
If it does, those c for which the t is relevant,
are the t's friend. And, the t is the c's marker.
For each row, we want to identify the column(s) that specifically prefer(s)
the row. We say that such a column is a friend (or the best friend if it is
the only) for the row.
The simplest example: imagine that only one column pays attention to our row.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/favorov/friends.test}
  \item \url{https://arxiv.org/abs/2512.24843}
  \item Report bugs at \url{https://github.com/favorov/friends.test/issues}
}

}
\author{
\strong{Maintainer}: Alexander Favorov \email{favorov@sensi.org} (\href{https://orcid.org/0000-0003-2595-1091}{ORCID})

Authors:
\itemize{
  \item Alexander Favorov \email{favorov@sensi.org} (\href{https://orcid.org/0000-0003-2595-1091}{ORCID})
  \item Alexandra Suvorikova \email{a.suvorikova@gmail.com} [contributor]
  \item Alexey Kroshnin \email{kroshnin@wias-berlin.de} [contributor]
  \item Dmitrijs Lvovs \email{dmitrijs.lvovs@gmail.com} [contributor]
  \item Vera Mukhina \email{atgc@vigg.ru} [contributor]
  \item Andrey A. Mironov \email{mironov@bioinf.fbb.msu.ru} [contributor]
  \item Ludmila Danilova \email{ludmila.danilova@gmail.com} [contributor]
}

}
\keyword{internal}
