% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/friends.test.cogaps.example.r
\docType{data}
\name{friends.test.cogaps.example}
\alias{friends.test.cogaps.example}
\title{An example of loadings of genes in CoGAPS - identified patterns.}
\format{
An object of class \code{list} of length 3.
}
\source{
\url{https://zenodo.org/records/7709664/files/cogapsresult.Rds}
}
\usage{
data(friends.test.cogaps.example)
}
\description{
An example of loadings of genes in CoGAPS - identified patterns.
}
\keyword{datasets}
